refactor(quantizer): optimize quantizer's compute
