+++ src/quantization/CMakeLists.txt
@@ -10,7 +10,6 @@ set (QUANTIZER_SRC
         product_quantization/pq_fastscan_quantizer.cpp
         product_quantization/product_quantizer.cpp
         rabitq_quantization/rabitq_quantizer.cpp
-        transform_quantization/transform_quantizer.cpp
         quantizer_parameter.cpp
         fp32_quantizer_parameter.cpp
         scalar_quantization/sq8_quantizer_parameter.cpp
@@ -24,7 +23,6 @@ set (QUANTIZER_SRC
         product_quantization/product_quantizer_parameter.cpp
         product_quantization/pq_fastscan_quantizer_parameter.cpp
         transform_quantization/transform_quantizer_parameter.cpp
-        quantizer_interface.cpp
 )
 
 add_library (quantizer OBJECT ${QUANTIZER_SRC})
