+++ src/quantization/computer.h
@@ -18,7 +18,6 @@
 #include <cstdint>
 #include <memory>
 
-#include "computer_interface.h"
 #include "metric_type.h"
 #include "vsag/allocator.h"
 
@@ -26,34 +25,38 @@ namespace vsag {
 
 using DataType = float;
 
-template <typename T>
-class Computer : public ComputerInterface, public std::enable_shared_from_this<Computer<T>> {
+class ComputerInterface {
 public:
-    ~Computer() override {
-        this->allocator_->Deallocate(buf_);
-    }
+    ComputerInterface() = default;
 
+    virtual ~ComputerInterface() = default;
+};
+
+using ComputerInterfacePtr = std::shared_ptr<ComputerInterface>;
+
+template <typename T>
+class Computer : public ComputerInterface {
+public:
     explicit Computer(const T* quantizer, Allocator* allocator)
         : quantizer_(quantizer), allocator_(allocator){};
 
-    ComputerInterfacePtr
-    GetComputerInterfacePtr() override {
-        return this->shared_from_this();
+    ~Computer() override {
+        quantizer_->ReleaseComputer(*this);
     }
 
     void
-    SetQuery(const DataType* query) override {
-        quantizer_->ProcessQuery(query, this->shared_from_this());
+    SetQuery(const DataType* query) {
+        quantizer_->ProcessQuery(query, *this);
     }
 
     inline void
     ComputeDist(const uint8_t* codes, float* dists) {
-        quantizer_->ComputeDist(this->shared_from_this(), codes, dists);
+        quantizer_->ComputeDist(*this, codes, dists);
     }
 
     inline void
     ScanBatchDists(uint64_t count, const uint8_t* codes, float* dists) {
-        quantizer_->ScanBatchDists(this->shared_from_this(), count, codes, dists);
+        quantizer_->ScanBatchDists(*this, count, codes, dists);
     }
 
     inline void
@@ -65,22 +68,65 @@ public:
                        float& dists2,
                        float& dists3,
                        float& dists4) {
-        quantizer_->ComputeDistsBatch4(this->shared_from_this(),
-                                       codes1,
-                                       codes2,
-                                       codes3,
-                                       codes4,
-                                       dists1,
-                                       dists2,
-                                       dists3,
-                                       dists4);
+        quantizer_->ComputeDistsBatch4(
+            *this, codes1, codes2, codes3, codes4, dists1, dists2, dists3, dists4);
     }
 
 public:
     Allocator* const allocator_{nullptr};
     const T* quantizer_{nullptr};
     uint8_t* buf_{nullptr};
-    ComputerInterfacePtr inner_computer_{nullptr};
+};
+
+template <typename QuantImpl, MetricType metric>
+class TransformQuantizer;
+
+template <typename QuantImpl, MetricType metric>
+class Computer<TransformQuantizer<QuantImpl, metric>> : public ComputerInterface {
+public:
+    explicit Computer(const TransformQuantizer<QuantImpl, metric>* quantizer, Allocator* allocator)
+        : quantizer_(quantizer), allocator_(allocator) {
+        inner_computer_ = new Computer<QuantImpl>(quantizer_->quantizer_.get(), allocator);
+    }
+
+    ~Computer() override {
+        quantizer_->ReleaseComputer(*this);
+        delete inner_computer_;
+    }
+
+    void
+    SetQuery(const DataType* query) {
+        quantizer_->ProcessQuery(query, *this);
+    }
+
+    inline void
+    ComputeDist(const uint8_t* codes, float* dists) {
+        quantizer_->ComputeDist(*this, codes, dists);
+    }
+
+    inline void
+    ScanBatchDists(uint64_t count, const uint8_t* codes, float* dists) {
+        quantizer_->ScanBatchDists(*this, count, codes, dists);
+    }
+
+    inline void
+    ComputeDistsBatch4(const uint8_t* codes1,
+                       const uint8_t* codes2,
+                       const uint8_t* codes3,
+                       const uint8_t* codes4,
+                       float& dists1,
+                       float& dists2,
+                       float& dists3,
+                       float& dists4) {
+        quantizer_->ComputeDistsBatch4(
+            *this, codes1, codes2, codes3, codes4, dists1, dists2, dists3, dists4);
+    }
+
+public:
+    Allocator* const allocator_{nullptr};
+    const TransformQuantizer<QuantImpl, metric>* quantizer_{nullptr};
+    uint8_t* buf_{nullptr};
+    Computer<QuantImpl>* inner_computer_{nullptr};
 };
 
 }  // namespace vsag
